Some Saves

+++ src/Util.cpp
@@ -85,7 +85,7 @@ std::vector<api::Entity> Util::pointsInRadius(std::vector<api::Entity> toCheck, 
 
 float Util::distance(api::Position2D p1, api::Position2D p2)
 {
-	return sqrt(pow(p2.x - p1.x, 2) + pow(p2.y - p1.y, 2));
+	return float(sqrt(pow(p2.x - p1.x, 2) + pow(p2.y - p1.y, 2)));
 }
 
 api::Position2D Util::A_B_Offsetter(api::Position2D A, api::Position2D B, float Range)
